partialize graphql to separate customizations from core
