+++ website/src/lib/wordpress/graphql/flexWysiwygContent.js
@@ -1,0 +1,14 @@
+import sectionOptions from './fragmentSectionOptions';
+
+export function flexWysiwygContent(template = 'Flex') {
+  const fragment = /* GraphQL */ `
+    ... on Template_${template}_Acfflex_FlexContent_WysiwygContent {
+      # contentWidth
+      sectionHeading
+      wysiwygContent
+      ${sectionOptions}
+    }
+  `;
+
+  return fragment;
+}
