+++ website/src/lib/wordpress/graphql/fragmentCategories.js
@@ -1,0 +1,11 @@
+const fragmentCategories = /* GraphQL */ `
+  categories {
+    nodes {
+      name
+      slug
+      uri
+    }
+  }
+`;
+
+export default fragmentCategories;
