+++ website/src/lib/wordpress/graphql/fragmentPageOptions.js
@@ -1,0 +1,12 @@
+const fragmentPageOptions = /* GraphQL */ `
+  acfPageOptions {
+    footerHideNav
+    footerHideSearch
+    footerHideSignup
+    footerHideSocial
+    footerStyle
+    headerHideNav
+  }
+`;
+
+export default fragmentPageOptions;
