+++ website/src/lib/wordpress/graphql/fragmentSectionOptions.js
@@ -1,0 +1,9 @@
+const sectionOptions = /* GraphQL */ `
+  backgroundColor
+  hideSection
+  fieldGroupName
+  sectionClasses
+  sectionSlug
+`;
+
+export default sectionOptions;
